+++ Firestore/Swift/Source/SwiftAPI/Pipeline/Expressions/FunctionExpressions/FunctionExpression.swift
@@ -24,6 +24,7 @@ public class FunctionExpression: Expression, BridgeWrapper, @unchecked Sendable 
 
   let functionName: String
   let args: [Expression]
+  let options: [String: Sendable]?
 
   /// The error message associated with this expression or its arguments, if any.
   var errorMessage: String? {
@@ -36,12 +37,17 @@ public class FunctionExpression: Expression, BridgeWrapper, @unchecked Sendable 
   /// - Parameters:
   ///   - functionName: The name of the function.
   ///   - args: The arguments to the function.
-  public init(functionName: String, args: [Expression]) {
+  ///   - options: The options for the function call.
+  public init(functionName: String,
+              args: [Expression],
+              options: [String: Sendable]? = nil) {
     self.functionName = functionName
     self.args = args
+    self.options = options
     bridge = FunctionExprBridge(
       name: functionName,
-      args: self.args.map { $0.toBridge() }
+      args: self.args.map { $0.toBridge() },
+      options: options?.mapValues { Helper.sendableToExpr($0).toBridge() }
     )
   }
 }