+++ Firestore/Swift/Tests/Integration/PipelineTests.swift
@@ -3038,19 +3038,10 @@ class PipelineIntegrationTests: FSTIntegrationTestCase {
 
     let snapshot = try await pipeline.execute()
 
-    let expectedResults: [[String: Sendable?]]
-    switch FSTIntegrationTestCase.targetBackend() {
-    case .nightly:
-      expectedResults = [
-        ["title": "The Hitchhiker's Guide to the Galaxy", "awards": ["hugo": true]],
-        ["title": "Dune", "awards": ["hugo": true]],
-      ]
-    default:
-      expectedResults = [
-        ["title": "The Hitchhiker's Guide to the Galaxy", "awards.hugo": 1],
-        ["title": "Dune", "awards.hugo": 1],
-      ]
-    }
+    let expectedResults: [[String: Sendable?]] = [
+      ["title": "The Hitchhiker's Guide to the Galaxy", "awards": ["hugo": true]],
+      ["title": "Dune", "awards": ["hugo": true]],
+    ]
 
     TestHelper.compare(snapshot: snapshot, expected: expectedResults, enforceOrder: true)
   }
@@ -3073,32 +3064,18 @@ class PipelineIntegrationTests: FSTIntegrationTestCase {
 
     XCTAssertEqual(snapshot.results.count, 2, "Should retrieve two documents")
 
-    let expectedResultsArray: [[String: Sendable?]]
-    switch FSTIntegrationTestCase.targetBackend() {
-    case .nightly:
-      expectedResultsArray = [
-        [
-          "title": "The Hitchhiker's Guide to the Galaxy",
-          "nestedField": ["level": [String: Any]()],
-          "nested": true,
-        ],
-        [
-          "title": "Dune",
-          "nestedField": ["level": [String: Any]()],
-          "nested": nil,
-        ],
-      ]
-    default:
-      expectedResultsArray = [
-        [
-          "title": "The Hitchhiker's Guide to the Galaxy",
-          "nested": true,
-        ],
-        [
-          "title": "Dune",
-        ],
-      ]
-    }
+    let expectedResultsArray: [[String: Sendable?]] = [
+      [
+        "title": "The Hitchhiker's Guide to the Galaxy",
+        "nestedField": ["level": [String: Any]()],
+        "nested": true,
+      ],
+      [
+        "title": "Dune",
+        "nestedField": ["level": [String: Any]()],
+        "nested": nil,
+      ],
+    ]
     TestHelper.compare(
       snapshot: snapshot,
       expected: expectedResultsArray,
@@ -3868,6 +3845,44 @@ class PipelineIntegrationTests: FSTIntegrationTestCase {
     )
   }
 
+  func testSupportsParent() async throws {
+    let db = firestore()
+    let randomCol = collectionRef()
+
+    // Add a dummy document directly to randomCol so the pipeline has input data to process
+    try await randomCol.document("dummy").setData(["dummy": 1])
+
+    let docRef = randomCol.document("book4").collection("reviews").document("review1")
+    try await docRef.setData(["foo": "bar"])
+
+    let pipeline = db.pipeline()
+      .collection(randomCol.path)
+      .limit(1)
+      .select([
+        Constant(docRef).parent().as("parentRefStatic"),
+        Constant(docRef).parent().as("parentRefInstance"),
+      ])
+      .select([
+        Field("parentRefStatic").documentId().as("parentIdStatic"),
+        Field("parentRefInstance").documentId().as("parentIdInstance"),
+      ])
+
+    let snapshot = try await pipeline.execute()
+
+    XCTAssertEqual(snapshot.results.count, 1)
+
+    let expectedResults: [String: Sendable] = [
+      "parentIdStatic": "book4",
+      "parentIdInstance": "book4",
+    ]
+
+    if let resultDoc = snapshot.results.first {
+      TestHelper.compare(pipelineResult: resultDoc, expected: expectedResults)
+    } else {
+      XCTFail("No document retrieved for parent test")
+    }
+  }
+
   func testSubstring() async throws {
     let collRef = collectionRef(withDocuments: TestHelper.bookDocs)
     let db = collRef.firestore
@@ -4043,6 +4058,210 @@ class PipelineIntegrationTests: FSTIntegrationTestCase {
       ]],
       enforceOrder: false
     )
+  }
+
+  func testLTrim() async throws {
+    let collRef = collectionRef(withDocuments: TestHelper.bookDocs)
+    let db = collRef.firestore
+
+    let pipeline = db.pipeline()
+      .collection(collRef.path)
+      .limit(1)
+      .replace(with: MapExpression([
+        "strWhitespace": "  trimMe  ",
+        "strChars": "xxxtrimMe",
+        "chars": "x",
+      ]))
+      .select([
+        Constant("  trimMe  ").ltrim().as("constLtrim"),
+        Field("strWhitespace").ltrim().as("fieldLtrim"),
+        Constant("xxxtrimMe").ltrim("x").as("constLtrimValuePrim"),
+        Constant("xxxtrimMe").ltrim(Field("chars")).as("constLtrimValueExpr"),
+        Field("strChars").ltrim("x").as("fieldLtrimValuePrim"),
+        Field("strChars").ltrim(Field("chars")).as("fieldLtrimValueExpr"),
+      ])
+
+    let snapshot = try await pipeline.execute()
+
+    let expectedResults: [[String: Sendable]] = [
+      [
+        "constLtrim": "trimMe  ",
+        "fieldLtrim": "trimMe  ",
+        "constLtrimValuePrim": "trimMe",
+        "constLtrimValueExpr": "trimMe",
+        "fieldLtrimValuePrim": "trimMe",
+        "fieldLtrimValueExpr": "trimMe",
+      ],
+    ]
+
+    TestHelper.compare(snapshot: snapshot, expected: expectedResults, enforceOrder: false)
+  }
+
+  func testRTrim() async throws {
+    let collRef = collectionRef(withDocuments: TestHelper.bookDocs)
+    let db = collRef.firestore
+
+    let pipeline = db.pipeline()
+      .collection(collRef.path)
+      .limit(1)
+      .replace(with: MapExpression([
+        "strWhitespace": "  trimMe  ",
+        "strChars": "trimMexxx",
+        "chars": "x",
+      ]))
+      .select([
+        Constant("  trimMe  ").rtrim().as("constRtrim"),
+        Field("strWhitespace").rtrim().as("fieldRtrim"),
+        Constant("trimMexxx").rtrim("x").as("constRtrimValuePrim"),
+        Constant("trimMexxx").rtrim(Field("chars")).as("constRtrimValueExpr"),
+        Field("strChars").rtrim("x").as("fieldRtrimValuePrim"),
+        Field("strChars").rtrim(Field("chars")).as("fieldRtrimValueExpr"),
+      ])
+
+    let snapshot = try await pipeline.execute()
+
+    let expectedResults: [[String: Sendable]] = [
+      [
+        "constRtrim": "  trimMe",
+        "fieldRtrim": "  trimMe",
+        "constRtrimValuePrim": "trimMe",
+        "constRtrimValueExpr": "trimMe",
+        "fieldRtrimValuePrim": "trimMe",
+        "fieldRtrimValueExpr": "trimMe",
+      ],
+    ]
+
+    TestHelper.compare(snapshot: snapshot, expected: expectedResults, enforceOrder: false)
+  }
+
+  func testStringRepeat() async throws {
+    let collRef = collectionRef(withDocuments: TestHelper.bookDocs)
+    let db = collRef.firestore
+
+    let pipeline = db.pipeline()
+      .collection(collRef.path)
+      .limit(1)
+      .replace(with: MapExpression([
+        "str": "ha",
+        "count": 3,
+      ]))
+      .select([
+        Constant("ha").stringRepeat(3).as("constPrim"),
+        Constant("ha").stringRepeat(Field("count")).as("constExpr"),
+        Field("str").stringRepeat(3).as("fieldPrim"),
+        Field("str").stringRepeat(Field("count")).as("fieldExpr"),
+      ])
+
+    let snapshot = try await pipeline.execute()
+
+    let expectedResults: [[String: Sendable]] = [
+      [
+        "constPrim": "hahaha",
+        "constExpr": "hahaha",
+        "fieldPrim": "hahaha",
+        "fieldExpr": "hahaha",
+      ],
+    ]
+
+    TestHelper.compare(snapshot: snapshot, expected: expectedResults, enforceOrder: false)
+  }
+
+  func testStringReplaceAll() async throws {
+    let collRef = collectionRef(withDocuments: TestHelper.bookDocs)
+    let db = collRef.firestore
+
+    let pipeline = db.pipeline()
+      .collection(collRef.path)
+      .limit(1)
+      .replace(with: MapExpression([
+        "str": "hello world",
+        "old": "o",
+        "new": "a",
+      ]))
+      .select([
+        Constant("hello world").stringReplaceAll("o", with: "a").as("constPrim"),
+        Constant("hello world").stringReplaceAll(Field("old"), with: Field("new")).as("constExpr"),
+        Field("str").stringReplaceAll("o", with: "a").as("fieldPrim"),
+        Field("str").stringReplaceAll(Field("old"), with: Field("new")).as("fieldExpr"),
+      ])
+
+    let snapshot = try await pipeline.execute()
+
+    let expectedResults: [[String: Sendable]] = [
+      [
+        "constPrim": "hella warld",
+        "constExpr": "hella warld",
+        "fieldPrim": "hella warld",
+        "fieldExpr": "hella warld",
+      ],
+    ]
+
+    TestHelper.compare(snapshot: snapshot, expected: expectedResults, enforceOrder: false)
+  }
+
+  func testStringReplaceOne() async throws {
+    let collRef = collectionRef(withDocuments: TestHelper.bookDocs)
+    let db = collRef.firestore
+
+    let pipeline = db.pipeline()
+      .collection(collRef.path)
+      .limit(1)
+      .replace(with: MapExpression([
+        "str": "hello world",
+        "old": "o",
+        "new": "a",
+      ]))
+      .select([
+        Constant("hello world").stringReplaceOne("o", with: "a").as("constPrim"),
+        Constant("hello world").stringReplaceOne(Field("old"), with: Field("new")).as("constExpr"),
+        Field("str").stringReplaceOne("o", with: "a").as("fieldPrim"),
+        Field("str").stringReplaceOne(Field("old"), with: Field("new")).as("fieldExpr"),
+      ])
+
+    let snapshot = try await pipeline.execute()
+
+    let expectedResults: [[String: Sendable]] = [
+      [
+        "constPrim": "hella world",
+        "constExpr": "hella world",
+        "fieldPrim": "hella world",
+        "fieldExpr": "hella world",
+      ],
+    ]
+
+    TestHelper.compare(snapshot: snapshot, expected: expectedResults, enforceOrder: false)
+  }
+
+  func testStringIndexOf() async throws {
+    let collRef = collectionRef(withDocuments: TestHelper.bookDocs)
+    let db = collRef.firestore
+
+    let pipeline = db.pipeline()
+      .collection(collRef.path)
+      .limit(1)
+      .replace(with: MapExpression([
+        "str": "hello world",
+        "sub": "world",
+      ]))
+      .select([
+        Constant("hello world").stringIndexOf("world").as("constPrim"),
+        Constant("hello world").stringIndexOf(Field("sub")).as("constExpr"),
+        Field("str").stringIndexOf("world").as("fieldPrim"),
+        Field("str").stringIndexOf(Field("sub")).as("fieldExpr"),
+      ])
+
+    let snapshot = try await pipeline.execute()
+
+    let expectedResults: [[String: Sendable]] = [
+      [
+        "constPrim": 6,
+        "constExpr": 6,
+        "fieldPrim": 6,
+        "fieldExpr": 6,
+      ],
+    ]
+
+    TestHelper.compare(snapshot: snapshot, expected: expectedResults, enforceOrder: false)
   }
 
   func testSplitWorks() async throws {