+++ Firestore/CHANGELOG.md
@@ -1,7 +1,12 @@
 # Unreleased
 - [changed] Drop C++14 support, Firestore SDK now requires at least C++17.
+
+# 12.12.0
+- [feature] Added support for the `parent` Pipeline expression. (#16010)
 - [feature] Added support for `timestampTruncate`, `timestampDiff`, and `timestampExtract` Pipeline expressions. (#15984)
 - [feature] Added support for Firestore Pipeline Subqueries and scope bridging with `Subcollection`, `define()`, `toArrayExpression()`, `toScalarExpression()`, `Variable`, and `CurrentDocument` APIs.
+- [feature] Added support for Pipeline expressions `ltrim`, `rtrim`, `stringIndexOf`, `stringReplaceOne`, `stringReplaceAll`, and `stringRepeat`.
+- [feature] Add preview support for full-text search and geo search.
 - [feature] Added support for Pipeline expressions `nor` and `switchOn`. (#15943)
 - [feature] Added support for Pipeline expressions `ifNull` and `coalesce`. (#15989)
 - [feature] Added support for Pipeline expression `isType`.